Integrate updateCards to dashboard_card_controller
